requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
idna==2.10
Pillow==8.1.0
pony==0.7.14
psycopg2==2.8.6
requests==2.25.0
six==1.15.0
soupsieve==2.2
urllib3==1.26.2
vk-api==11.9.1
